define migration test for advanced cluster using single and multi cloud provider configurations
